# doc/api.yaml
openapi: 3.0.0
info:
  description: WASAText app backend REST API
  title: WASAText API
  version: 1.0.0
servers:
  - url: http://localhost:3000/v1
...
security:
  - bearerAuth: []
paths:
...
  /session/login:
    post:
      tags: ["Login"]
      summary: Logs in the user
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.  
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              type: object
              description: "Represents a user's name, with validation rules for length and format."
              properties:
                name:
                  type: string
                  description: "The user's name, which must be between 3 and 16 characters in length and match the specified pattern."
                  example: Maria
                  pattern: "^.*?$"
                  minLength: 3
                  maxLength: 16
        required: true
      responses:
        "201":
          description: User log-in action successful
          content:
            application/json:
              schema:
                type: object
                description: "Represents a unique identifier, typically used to reference an entity or object in the system."
                properties:
                  identifier:
                    type: string
                    description: "A unique identifier for an entity, often used to retrieve or reference specific objects."
                    example: "abcdef012345"
                    minLength: 12
                    maxLength: 12
                    pattern: "^[a-zA-Z0-9]{12}$"
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  ##getUserProfile
  /users/{username}/profile:
    get:
      parameters:
      - $ref: "#/components/parameters/username"
      tags: ["User"]
      summary: "Get profile data of a user."
      description: "Retrieve a user's profile data by username."
      operationId: getUserProfile
      responses:
        "200":
          description: "Profile loaded."
          content:
            application/json:
              schema:
                type: object
                description: "Represents user profile information, including the user’s handle and display name."
                properties:
                  user_handle:
                    type: string
                    description: "The unique handle or username of the user."
                    example: "fiore_rossi"
                    minLength: 3
                    maxLength: 30
                    pattern: "^[a-zA-Z0-9_]{3,30}$"
                  screen_name:
                    type: string
                    description: "The display name of the user, typically shown on their profile."
                    example: "Fiore Rossi"
                    minLength: 2
                    maxLength: 50
                    pattern: "^[a-zA-Z ]{2,50}$"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalServerError" } 
        
  ##getUserPicture
  /users/{username}/picture:
    parameters:
      - $ref: "#/components/parameters/username"
    get:
      tags: ["Profile picture"]
      summary: "Get profile picture of a user."
      description: "Get a user's profile picture."
      operationId: getUserPicture
      responses:
        "200":
          description: "Picture loaded."
          content:
            image/jpeg:
              schema:
                type: string
                description: "Represents file uploads for the profile picture."
                format: binary
                ##minLength: 10
                ##maxLength: 100
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  ##setMyPhoto
    put:
      tags: ["Profile picture"]
      summary: "Upload profile picture."
      description: |-
        The user uploads a profile picture.
      operationId: setMyPhoto
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: "Represents a request to upload a picture, typically used for profile or group images."
              properties:
                photo:
                  type: string
                  format: binary
                  description: "The picture to upload."
                  ##minLength: 10
                  ##maxLength: 100
      responses:
        "200":
          description: "Picture uploaded."
          content:
            application/json:
              schema:
                type: object
                description: "Represents the response after successfully uploading a profile picture, including the confirmation message and the URL of the uploaded image."
                properties:
                  message:
                    type: string
                    description: "A message confirming the successful upload of the profile picture."
                    example: "Picture uploaded successfully."
                    minLength: 10
                    maxLength: 100
                    pattern: "^[a-zA-Z0-9 .,!?']+$"

                  profile_picture_url:
                    type: string
                    description: "The URL where the uploaded profile picture can be accessed."
                    format: url
                    example: "https://example.com/profiles/user_uploaded.jpg"
                    minLength: 15
                    maxLength: 255
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  ##setMyUserName
  /users/{username}:
    put:
      parameters:
      - $ref: "#/components/parameters/username"
      tags: ["User"]
      summary: "Update username."
      description: |-
        Updates the username for the user unless such a username is already taken
        by another user.
      operationId: setMyUserName
      requestBody:
        description: "Username"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "200":
          description: "Username updated successfully"
          content:
            application/json:
              schema: 
                type: object
                description: "Represents a request to set or update a user's handle (username)."
                properties:
                  user_handle:
                    type: string
                    description: "The updated username for the user."
                    example: "new_username"
                    minLength: 3
                    maxLength: 30
                    pattern: "^[a-zA-Z0-9_]{3,30}$"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: "Username already taken"
          content:
            application/json:
              schema:
                type: object
                description: "Represents a response indicating that the username is already taken and cannot be used."
                properties:
                  message:
                    type: string
                    description: "A message informing the user that the selected username is already in use."
                    example: "This username is already taken. Please choose a different one."
                    minLength: 10
                    maxLength: 100
                    pattern: "^[a-zA-Z0-9 .,!?']+$"
        "500": { $ref: "#/components/responses/InternalServerError" }

  ##getMyConversations
...
  /users/{username}/conversations/{conversation_id}:
    parameters:
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/conversation_id"
    get:
      tags: ["Conversation"]
      summary: "Get a specific conversation."
      description: "Get a user's specific conversations. Only the user can access their conversations."
      operationId: getConversation
      responses:
        "200":
          description: "Conversation details retrieved successfully."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  ##sendMessage
  /users/{username}/conversations/{conversation_id}/messages/:
    parameters:
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/conversation_id"
    post:
      tags: ["Message"]
      summary: "Send a new message."
      description: "Send a message from the logged-in user to another user"
      operationId: sendMessage
      requestBody:
        description: "The content of the message to be sent"
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Represents the structure of a message, including its type and content."
              properties:
                type:
                  type: string
                  description: "Type of message content (e.g., 'text' or 'image')."
                  example: "text"
                  minLength: 3
                  maxLength: 20
                  pattern: "^[a-zA-Z]+$"

                content:
                  type: string
                  description: "Message text or image URL, depending on the type."
                  example: "Hey there."
                  minLength: 1
                  maxLength: 500
                  pattern: "^[a-zA-Z0-9 .,!?'\n]*$" 
              required:
                - type
                - content
      responses:
        "201":
          description: "Message sent."
          content:
            application/json:
              schema:
                type: object
                description: "Represents a response confirming the successful sending of a message."
                properties:
                  message:
                    type: string
                    description: "A message confirming the successful delivery of the message."
                    example: "Message sent successfully."
                    minLength: 10
                    maxLength: 100
                    pattern: "^[a-zA-Z0-9 .,!?']+$"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  ##forwardMessage
  /users/{username}/conversations/{conversation_id}/messages/{message_id}/forward:
    parameters:
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/conversation_id"
      - $ref: "#/components/parameters/message_id"
    post:
      tags: ["Message"]
      summary: "Forward a message."
      description: "Forward a message from one user to another."
      operationId: forwardMessage
      requestBody:
        description: "The target conversation or recipient for the forwarded message."
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Represents the data required to forward a message to another conversation or user."
              properties:
                target_conversation_id:
                  type: string
                  description: "ID of the conversation to forward the message to."
                  example: "5678"
                  minLength: 1
                  maxLength: 50
                  pattern: "^[0-9]+$"  

                recipient_username:
                  type: string
                  description: "Username of the recipient if forwarding directly to another user."
                  example: "user_name"
                  minLength: 3
                  maxLength: 30
                  pattern: "^[a-zA-Z0-9_]{3,30}$" 
              required:
                - target_conversation_id
      responses:
        "200":
          description: "Message forwarded successfully."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  ##deleteMessage
  /users/{username}/conversations/{conversation_id}/messages/{message_id}:
    parameters:
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/conversation_id"
      - $ref: "#/components/parameters/message_id"
    delete:
      tags: ["Message"]
      summary: "Delete a message."
      description: "Delete a specific message from a conversation."
      operationId: deleteMessage
      responses:
        "204":
          description: "Message deleted successfully."
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  ##commentMessage
...
components:
  schemas:
    User:
      title: User
      description: "This object represents a single User."
      type: object
      properties:
        current_username:
          description: "Alias for the User."
          type: string
          pattern: '^[^\s]+$'
          minLength: 3
          maxLength: 16
          example: "slow_koala"
          readOnly: false
        id:
          description: "Unique identifier for each user that allows username changing."
          type: integer
          minimum: 0 
          maximum: 9999999
          example: 1
          readOnly: true
      required:
        - current_username
        - id
    Conversation:
      type: object
      description: "Represents a conversation between participants, including its metadata."
      properties:
        conversation_id:
          type: string
          description: "Unique identifier for the conversation."
          example: "abc123"
          minLength: 3
          maxLength: 50
          pattern: "^[a-zA-Z0-9_]+$"  
        participants:
          type: array
          description: "A list of participants in the conversation, represented by their usernames."
          items:
            type: string
            description: "Username of a participant in the conversation."
            minLength: 3
            maxLength: 30
            pattern: "^[a-zA-Z0-9_ ]+$"
          example: ["Fiore Rossi", "Ale Bianchi"]
          minItems: 2  
          maxItems: 50
        last_message:
          type: string
          description: "Content of the last message in the conversation."
          example: "Hi how are you doing"
          minLength: 1  
          maxLength: 500  
          pattern: "^[a-zA-Z0-9 ]+$"
      required:
        - participants
    Message:
      title: Message
      description: "This object represents a single message."
      type: object
      properties:
        id:
          description: "Unique identifier for each user that allows username changing."
          type: integer
          minimum: 0 
          maximum: 9999999
          example: 1
          readOnly: true
        timestamp:
          description: "Timestamp when the message was sent."
          type: string
          format: date-time
          example: "2023-10-19T15:23:00Z"
          readOnly: true
          minLength: 20  
          maxLength: 30
        preview:
          description: "A preview of the message, either a text snippet for text messages or a thumbnail URL for image messages."
          type: object
          properties:
            type:
                description: "Type of preview, either 'text' or 'image'."
                type: string
                enum: ["text", "image"]
                example: "text"
                minLength: 3
                maxLength: 10
            content:
                description: "The actual preview content. This is either a text snippet or a URL to a thumbnail image."
                type: string
                example: "This is a preview of the message content."
                minLength: 1
                maxLength: 500
                pattern: "^[a-zA-Z0-9 .]+$"
            thumbnail_url:
                description: "URL of the thumbnail image if the message is an image."
                type: string
                format: url
                example: "https://example.com/path/to/thumbnail.jpg"
                minLength: 10
                maxLength: 200
          required:
          - type
          - content
        comments:
          description: "List of emoji reactions to the message."
          type: array
          items:
            type: object
            properties:
              emoji:
                description: "The emoji used to react to the message."
                type: string
                example: "👍"
                minLength: 1
                maxLength: 1
                pattern: "^[\u0000-\uFFFF]+$"
              user_id:
                description: "Unique identifier of the user who added the reaction."
                type: integer
                minimum: 0 
                maximum: 9999999
                example: 42
              timestamp:
                description: "Timestamp when the reaction was added."
                type: string
                format: date-time
                example: "2023-10-19T15:23:00Z"
                minLength: 20
                maxLength: 30
                readOnly: true
            required:
              - emoji
              - user_id
              - timestamp
          minItems: 1
          maxItems: 50
        message_status:
          description: "Indicates the sender’s username if the message was received, or checkmarks for sent message status."
          type: object
          oneOf:
              - properties:
                    type:
                        description: "Specifies that the message was received by the user."
                        type: string
                        enum: ["received"]
                    sender_username:
                        description: "Username of the sender if the message was received by the user."
                        type: string
                        example: "slow_koala"
                        minLength: 3
                        maxLength: 30
                        pattern: "^[A-Za-z0-9_]*$"
                required:
                    - type
                    - sender_username
              - properties:
                    type:
                        description: "Specifies that the message was sent by the user."
                        type: string
                        enum: ["sent"]
                    checkmarks:
                        description: "Status of the sent message. One checkmark for 'sent', two for 'delivered'."
                        type: integer
                        minimum: 0 
                        maximum: 9999999
                        enum: [1, 2]
                        example: 2
                required:
                    - type
                    - checkmarks
        message_content:
          description: "The content of the message, which can be either text or an image."
          type: object
          properties:
            type:
                description: "Type of content, either 'text' or 'image'."
                type: string
                enum: ["text", "image"]
                example: "text"
            text:
                description: "The text content of the message, present only if the type is 'text'."
                type: string
                example: "Hey bestie."
                minLength: 1
                maxLength: 500
                pattern: "^[A-Za-z0-9 !@#\\$%\\^&\\*()\\-=_+\\[\\]{};':\"\\\\|,.<>/?`~]*$"
            image_url:
                description: "URL of the image if the message contains an image."
                type: string
                format: url
                example: "https://example.com/path/to/image.jpg"
                minLength: 5
                maxLength: 2048
                pattern: "^https?://[\\w.-]+(?:\\.[\\w.-]+)+(?:/[\\w._~:/?#[\\]@!$&'()*+,;=%-]*)?$"
          required:
            - type
      required:
      - id
      - timestamp
      - preview
      - comments
      - message_content
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: "Invalid/malformed request."
      content:
        application/json:
          schema:
            type: object
            description: "An error response indicating the request was malformed or invalid."
            properties:
              message:
                type: string
                description: "An error response indicating the request was malformed or invalid."
                example: "The request body is malformed."
                minLength: 10
                maxLength: 200
                pattern: "^[A-Z].*\\.$"
    Unauthorized:
      description: "Authentication required."
      content:
        application/json:
          schema:
            type: object
            description: "An error response indicating that the user needs to authenticate."
            properties:
              message:
                type: string
                description: "An error response indicating that the user needs to authenticate."
                example: "Authentication is required to access this resource."
                minLength: 10
                maxLength: 200
                pattern: "^[A-Z].*\\.$"
    InternalServerError:
      description: "Server currently unavailable."
      content:
        application/json:
          schema:
            type: object
            description: "An error response indicating an internal server error."
            properties:
              message:
                type: string
                description: "An error response indicating an internal server error."
                example: "The server is currently unavailable, please try again later."
                minLength: 10
                maxLength: 200
                pattern: "^[A-Z].*\\.$"
    NotFound:
      description: "Resource not found."
      content:
        application/json:
          schema:
            type: object
            description: "An error response indicating the requested resource could not be found."
            properties:
              message:
                type: string
                description: "An error response indicating the requested resource could not be found."
                example: "The requested resource could not be found."
                minLength: 10
                maxLength: 200
                pattern: "^[A-Z].*\\.$"
    Forbidden:
      description: "The operation is forbidden."
      content:
        application/json:
          schema:
            type: object
            description: "An error response indicating that the operation is not allowed."
            properties:
              message:
                type: string
                description: "An error response indicating that the operation is not allowed."
                example: "You do not have permission to perform this operation."
                minLength: 10
                maxLength: 200
                pattern: "^[A-Z].*\\.$"
  parameters:
    username:
      name: username
      in: path
      required: true
      description: "Alias for the User."
      schema:
        type: string
        minLength: 6
        maxLength: 30
        pattern: "^[a-zA-Z0-9_]{6,30}$"
        example: booklover
    conversation_id:
      name: conversation_id
      in: path
      required: true
      description: "ID for a conversation."
      schema:
        type: string
        minLength: 6
        maxLength: 30
        example: convo123
        pattern: "^[a-zA-Z0-9_]+$"
    message_id:
      name: message_id
      in: path
      required: true
      description: "ID for a message."
      schema:
        type: string
        minLength: 6
        maxLength: 30
        example: message123
        pattern: "^[a-zA-Z0-9_]+$"
...